# api/docs/swagger.yaml
definitions:
...
  models.Category:
    properties:
      created_at:
        type: string
      id:
        type: string
      name:
        type: string
      parent_id:
        type: string
      updated_at:
        type: string
    type: object
  models.CategoryCU:
    properties:
      name:
        type: string
      parent_id:
        type: string
    type: object
  models.Error:
    properties:
      message:
        type: string
    type: object
  models.JustOrder:
    properties:
      book_id:
        type: string
      description:
        type: string
      id:
        type: string
    type: object
  models.Order:
    properties:
      book_id:
        type: string
      created_at:
        type: string
      description:
        type: string
      id:
        type: string
      updated_at:
        type: string
    type: object
  models.StandardErrorModel:
    properties:
      error:
        $ref: '#/definitions/models.Error'
    type: object
info:
  contact: {}
paths:
...
  /v1/categories/:
    get:
      consumes:
      - application/json
      description: This API for getting list of categories
      parameters:
      - description: Page
        in: query
        name: page
        type: string
      - description: Limit
        in: query
        name: limit
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Category'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.StandardErrorModel'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.StandardErrorModel'
      summary: ListCategories
      tags:
      - category
    post:
      consumes:
      - application/json
      description: This API for creating a new category
      parameters:
      - description: CategoryCreateRequest
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.CategoryCU'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Category'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.StandardErrorModel'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.StandardErrorModel'
      summary: CreateCategory
      tags:
      - category
  /v1/categories/{id}:
    delete:
      consumes:
      - application/json
      description: This API for deleting category
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: ""
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.StandardErrorModel'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.StandardErrorModel'
      summary: DeleteCategory
      tags:
      - category
    get:
      consumes:
      - application/json
      description: This API for getting category detail
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Category'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.StandardErrorModel'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.StandardErrorModel'
      summary: GetCategory
      tags:
      - category
    put:
      consumes:
      - application/json
      description: This API for updating category
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: string
      - description: categoryUpdateRequest
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.CategoryCU'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.CategoryCU'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.StandardErrorModel'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.StandardErrorModel'
      summary: UpdateCategory
      tags:
      - category
  /v1/orders/:
    get:
      consumes:
      - application/json
      description: This API for getting list of orders
      parameters:
      - description: Page
        in: query
        name: page
        type: string
      - description: Limit
        in: query
        name: limit
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Order'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.StandardErrorModel'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.StandardErrorModel'
      summary: ListOrders
      tags:
      - order
    post:
      consumes:
      - application/json
      description: This API for creating a new order
      parameters:
      - description: orderCreateRequest
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.JustOrder'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Order'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.StandardErrorModel'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.StandardErrorModel'
      summary: CreateOrder
      tags:
      - order
  /v1/orders/{id}:
    delete:
      consumes:
      - application/json
      description: This API for deleting order
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: ""
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.StandardErrorModel'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.StandardErrorModel'
      summary: DeleteOrder
      tags:
      - order
    get:
      consumes:
      - application/json
      description: This API for getting order detail
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Order'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.StandardErrorModel'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.StandardErrorModel'
      summary: GetOrder
      tags:
      - order
    put:
      consumes:
      - application/json
      description: This API for updating order
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: string
      - description: orderUpdateRequest
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.JustOrder'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Order'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.StandardErrorModel'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.StandardErrorModel'
      summary: UpdateOrder
      tags:
      - order
swagger: "2.0"